# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gym API",
        "description": "API pour la gestion de salle de sport",
        "contact": {
            "email": "admin@gym-gestion.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/subscription-status/list": {
            "get": {
                "tags": [
                    "Subscription status"
                ],
                "summary": "Liste des statuts de souscription",
                "operationId": "4170f3874aea5d9b19fa15dbba081210",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/subscription-status/{id}": {
            "get": {
                "tags": [
                    "Subscription status"
                ],
                "summary": "Détails d'un statut d'abonnement",
                "operationId": "ffc3578677d4f47fb2449577a87d3d13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du statut d'abonnement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/subscription-status/new": {
            "post": {
                "tags": [
                    "Subscription status"
                ],
                "summary": "Ajout d'un statut d'abonnement",
                "operationId": "0dad095013c54220da7a8e75d0fb0677",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "subscription_status": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "Active"
                                    },
                                    "active_flag": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Enregistrement effectué"
                    },
                    "401": {
                        "description": "Action non autorisée"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/subscription-status/status/{id}": {
            "patch": {
                "tags": [
                    "Subscription status"
                ],
                "summary": "Toggle active flag of a subscription status",
                "operationId": "67dd7193294e8025c185e57ddd9e63e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du statut d'abonnement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enregistrement effectué"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/subscription-status/update/{id}": {
            "put": {
                "tags": [
                    "Subscription status"
                ],
                "summary": "Mettre à jour un statut d'abonnement",
                "operationId": "c00e32c76272cf69a46efe03444ca884",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du statut d'abonnement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "subscription_status": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "Inactive"
                                    },
                                    "active_flag": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mis à jour effectuée avec succès"
                    },
                    "400": {
                        "description": "Demande invalide"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/subscription-status/delete/{id}": {
            "delete": {
                "tags": [
                    "Subscription status"
                ],
                "summary": "Suppression d'un statut d'abonnement",
                "operationId": "9d554e9281fa1969748ecccc2ede9d09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du statut d'abonnement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}